+++ NpgsqlRest/Exceptions/NpgsqlToHttpException.cs
@@ -2,8 +2,9 @@ using Npgsql;
 
 namespace NpgsqlRest;
 
-public class NpgsqlToHttpException(ErrorCodeMappingOptions entry, PostgresException innerException)
-    : Exception(entry.Title ?? innerException.MessageText, innerException)
+public class NpgsqlToHttpException(ErrorCodeMappingOptions entry, NpgsqlException innerException)
+    : Exception(
+        entry.Title ?? (innerException is PostgresException exception ? exception.MessageText : innerException.Message), innerException)
 {
     public ErrorCodeMappingOptions Mapping { get; } = entry;
     public string? SqlState { get; } = innerException.SqlState;